refactor/datatypes: move shared code from sequence
