+++ src/lib.rs
@@ -64,17 +64,14 @@ mod mutable_data;
 mod public_key;
 mod request;
 mod response;
+mod shared_data;
 mod transaction;
 mod utils;
 
 pub use append_only_data::{
-    Action as ADataAction, Address as ADataAddress, AppendOperation as ADataAppend, Data as AData,
-    Entries as ADataEntries, Entry as ADataEntry, ExpectedIndices as ADataIndices,
-    Index as ADataIndex, Kind as ADataKind, Owner as ADataOwner,
-    PrivatePermissionSet as ADataUnpubPermissionSet, PrivatePermissions as ADataUnpubPermissions,
-    PrivateSentriedSequence, PrivateSequence, PublicPermissionSet as ADataPubPermissionSet,
-    PublicPermissions as ADataPubPermissions, PublicSentriedSequence, PublicSequence, Sequence,
-    SequencePermissions as ADataPermissions, User as ADataUser,
+    AppendOperation as ADataAppend, Data as AData, Entries as ADataEntries, Entry as ADataEntry,
+    PrivateSentriedSequence, PrivateSequence, PublicSentriedSequence, PublicSequence, Sequence,
+    SequencePermissions as ADataPermissions,
 };
 pub use coins::{Coins, MAX_COINS_VALUE};
 pub use errors::{EntryError, Error, Result};
@@ -101,6 +98,10 @@ pub use public_key::{PublicKey, Signature};
 pub use request::{LoginPacket, Request, MAX_LOGIN_PACKET_BYTES};
 pub use response::Response;
 pub use sha3::Sha3_512 as Ed25519Digest;
+pub use shared_data::{
+    Action as ADataAction, Address, ExpectedIndices, Index, Kind, Owner, PrivatePermissionSet,
+    PrivatePermissions, PublicPermissionSet, PublicPermissions, User,
+};
 pub use transaction::{Transaction, TransactionId};
 pub use utils::verify_signature;
